search by location complete
