+++ react-ui/src/components/Home-page/Home-page.css
@@ -31,23 +31,14 @@
     font-weight: 700; 
     line-height: 15.4px; 
   } 
-  
-  p { 
-    font-family: Bangers; 
-    font-size: 14px; 
-    font-style: normal; 
-    font-variant: normal; 
-    font-weight: 400; 
-    line-height: 20px; 
-  } 
 
   .searchBar {
     display: flex;
-    min-height: 100vh;
     flex-direction: column;
     align-items: center;
     font-size: calc(10px + 2vmin);
     color: black;
+    margin-bottom: 80px;
   }
 
   .locationBar {
